requirements.txt
Django>=1.5.1
django_extensions>=1.1.1
django-forms-builder>=0.8.4
South>=0.8.1
django_annoying>=0.7.7
